rename droidlet_nsp_model_wrapper -> nsp_querier
