+++ agents/craftassist/tests/test_nsp_wrapper_via_agent.py
@@ -6,7 +6,7 @@ import unittest
 from droidlet.dialog.dialogue_manager import DialogueManager
 from droidlet.memory.dialogue_stack import DialogueStack
 from droidlet.dialog.map_to_dialogue_object import DialogueObjectMapper
-from droidlet.perception.semantic_parsing.droidlet_nsp_model_wrapper import NSPQuerier
+from droidlet.perception.semantic_parsing.nsp_querier import NSPQuerier
 from agents.loco_mc_agent import LocoMCAgent
 from droidlet.shared_data_structs import MockOpt
